+++ src/game/Trail.java
@@ -14,7 +14,7 @@ public class Trail extends GameObject {
     private int width, height;
     private float life;
 
-    public Trail(int x, int y, ID id, Handler handler, Color color, int width, int height, float life) {
+    public Trail(float x, float y, ID id, Handler handler, Color color, int width, int height, float life) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -28,7 +28,7 @@ public class Trail extends GameObject {
 
         g2d.setComposite(makeTransparent(alpha));
         g.setColor(color);
-        g.fillRect(x, y, width, height);
+        g.fillRect((int)x, (int)y, width, height);
         g2d.setComposite(makeTransparent(1));
         
     }
